--- src/elem/Buttons.jsx
@@ -1,1 +1,0 @@
-import React from "react";
