--- src/elem/Input.jsx
@@ -1,1 +1,0 @@
-import React from "react";
