+++ src/page/Main.jsx
@@ -6,42 +6,43 @@ import NavigationBar from "../elem/NavigationBar";
 import RepPost from "../components/mainComponents/RepPost";
 import HotPosts from "../components/mainComponents/HotPosts";
 import AllPosts from "../components/mainComponents/AllPosts";
-import { useNavigate } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import { __getUsers } from "../redux/async/signup";
 import { __getRepresentative } from "../redux/async/upload";
 import { __getHotPosts } from "../redux/async/rank";
-import _ from "lodash";
 import { getCookie } from "../shared/cookie";
 import jwt from "jwt-decode"; // to get userId from loggedIn user's token
 
 const upButton = "/images/upArrow.png";
 
 const Main = (props) => {
-  const [scrollHeightInfo, SetScrollHeightInfo] = useState(0);
-
   const dispatch = useDispatch();
-  const navigate = useNavigate();
+
+  // userId 조회 by token
+  const token = getCookie("token");
+  const { userId } = jwt(token);
+
+  // 유저정보를 불러와서 토큰이 없다면 다시 로그인
+  // 유저정보 조회해서 프로필 사진 확보
+  useEffect(() => {
+    dispatch(__getUsers(userId));
+    dispatch(__getRepresentative(userId));
+    dispatch(__getHotPosts());
+  }, [dispatch]);
+
   // 유저정보 조회
   const userStatus = useSelector((state) => state.signup.userStatus);
   // 대표게시물 조회
   const repPost = useSelector((state) => state.upload.representative);
   // 랭크게시물 불러옴
   const hotPosts = useSelector((state) => state.rank.hotPosts);
-  // console.log(hotPosts);
 
-  const token = getCookie("token");
-  const { userId } = jwt(token);
+  // 유저 프로필 이미지 없을때 미리보기 이미지
   const preview_URL =
     "https://cdn.discordapp.com/attachments/1014169130045292625/1014194232250077264/Artboard_1.png";
-  // profile_pic를 정하는 부분
-  // const [profile, setProfile] = useState({
-  //   image_file: "",
-  //   preview_URL:
-  //     "https://cdn.discordapp.com/attachments/1014169130045292625/1014194232250077264/Artboard_1.png",
-  // });
 
   // toTop버튼
+  const [scrollHeightInfo, SetScrollHeightInfo] = useState(0);
   const showTopButton = () => {
     if (scrollHeightInfo > 2000) {
       //2000px밑으로 스크롤 내려갔을때 위로가는 Top 버튼 보이기
@@ -51,34 +52,13 @@ const Main = (props) => {
     }
   };
 
+  // 실행시 맨위로 올라옴
   const ScrollToTop = () => {
     window.scrollTo({
       top: 0,
       behavior: "smooth",
     });
   };
-
-  //스크롤 위치계산시 연산 너무 많이되는 것
-  //방지하기 위해 300ms 쓰로틀적용
-  const _scrollPosition = _.throttle(() => {
-    const scrollHeight = document.documentElement.scrollTop;
-    SetScrollHeightInfo(scrollHeight);
-  }, 300);
-
-  useEffect(() => {
-    window.addEventListener("scroll", _scrollPosition); // scroll event listener 등록
-    return () => {
-      window.removeEventListener("scroll", _scrollPosition); // scroll event listener 해제(스크롤이벤트 클린업)
-    };
-  }, [scrollHeightInfo]);
-
-  // 유저정보를 불러와서 토큰이 없다면 다시 로그인
-  // 유저정보 조회해서 프로필 사진 확보
-  useEffect(() => {
-    dispatch(__getUsers(userId));
-    dispatch(__getRepresentative(userId));
-    dispatch(__getHotPosts());
-  }, []);
 
   return (
     <Fragment>
@@ -99,11 +79,6 @@ const Main = (props) => {
             ) : (
               <Img url={userStatus?.imgUrl}></Img>
             )}
-            {/* <Img
-              url={
-                profile.image_file ? profile.image_file : profile.preview_URL
-              }
-            ></Img> */}
             {/* 대표게시물 출력 */}
             <RepPost myRepPost={repPost} />
             {/* 랭킹게시물 출력 */}
